Put index_schema.go into table_schema.go
